LayoutTests/webrtc/peer-connection-audio-mute.html

<!DOCTYPE html>
<html>
<head>
    <meta charset="utf-8">
    <title>Testing local audio capture playback causes "playing" event to fire</title>
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
</head>
<body>
    <script src ="routines.js"></script>
    <script>
    promise_test((test) => {
        if (window.testRunner)
            testRunner.setUserMediaPermission(true);

        return navigator.mediaDevices.getUserMedia({audio: true}).then((stream) => {
            if (window.internals)
                internals.useMockRTCPeerConnectionFactory("TwoRealPeerConnections");

            var stream;
            return new Promise((resolve, reject) => {
                createConnections((firstConnection) => {
                    firstConnection.addStream(stream);
                }, (secondConnection) => {
                    secondConnection.onaddstream = (streamEvent) => {
                        stream = streamEvent.stream;
                        resolve();
                    };
                });
            }).then(() => {
                return waitFor(500);
            }).then(() => {
                return analyseAudio(stream, 500).then((results) => {
                    assert_true(results.heardHum, "heard hum");
                    assert_true(results.heardBip, "heard bip");
                    assert_true(results.heardBop, "heard bop");
                });
            }).then(() => {
                stream.getAudioTracks().forEach((track) => {
                    track.enabled = false;
                });
                return waitFor(500);
            }).then(() => {
                return analyseAudio(stream, 500).then((results) => {
                    assert_false(results.heardHum, "heard hum");
                    assert_false(results.heardBip, "heard bip");
                    assert_false(results.heardBop, "heard bop");
                });
            }).then(() => {
                stream.getAudioTracks().forEach((track) => {
                    track.enabled = true;
                });
                return waitFor(500);
            }).then(() => {
                return analyseAudio(stream, 500).then((results) => {
                    assert_true(results.heardHum, "heard hum");
                    assert_true(results.heardBip, "heard bip");
                    assert_true(results.heardBop, "heard bop");
                });
            });
        });
    }, "Muting and unmuting an audio track");
    </script>
</body>
</html>
